# docs/swagger.yaml
definitions:
  Event:
    description: Fields untuk event
    properties:
      created_at:
        description: waktu pembuatan event
        type: string
      description:
        description: deskripsi event
        type: string
      event_date:
        description: tanggal event
        type: string
      event_type_id:
        description: id jenis event
        type: integer
      id:
        description: Id dari event
        type: integer
      location:
        description: lokasi event
        type: string
      modified_at:
        description: waktu terakhir event diubah
        type: string
      status:
        description: status event (open/closed)
        enum:
        - open
        - closed
        type: string
      title:
        description: nama event
        type: string
      user_id:
        description: Id user pembuat event
        type: integer
    required:
    - event_date
    - event_type_id
    - location
    - status
    - title
    type: object
  EventType:
    description: Fields untuk jenis event
    properties:
      created_at:
        description: Waktu pembuatan jenis event
        type: string
      id:
        description: Id dari jenis event
        type: integer
      modified_at:
        description: Waktu terakhir jenis event diubah
        type: string
      name:
        description: Nama jenis event
        type: string
    required:
    - name
    type: object
...
  ResponseError:
    description: menampilkan response error dengan pesan dan detail, bisa berupa error
      dari binding atau error lainnya jika detail tidak ada, maka akan diabaikan
    properties:
      detail: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  ResponseSuccess:
    description: menampilkan response sukses hanya dengan pesan
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
...
  ResponseSuccessWithData:
    description: menampilkan response sukses dengan pesan dan data bisa berupa struct,
      slice, atau map jika data tidak ada, maka akan diabaikan
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
info:
  contact: {}
paths:
  /api/event-types/:id/delete:
    delete:
      consumes:
      - application/json
      description: |-
        Menghapus jenis event
        * Hanya bisa diakses oleh admin
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil menghapus event type
          schema:
            $ref: '#/definitions/ResponseSuccess'
        "400":
          description: Gagal menghapus event type
          schema:
            $ref: '#/definitions/ResponseError'
      summary: Delete Event Type
      tags:
      - Event Type
  /api/event-types/:id/events:
    get:
      consumes:
      - application/json
      description: Mendapatkan daftar event berdasarkan jenis event
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil mendapatkan list event
          schema:
            $ref: '#/definitions/ResponseSuccessWithData'
        "400":
          description: Gagal mendapatkan list event
          schema:
            $ref: '#/definitions/ResponseError'
      summary: Get Event By Event Type Id
      tags:
      - Event Type
  /api/event-types/:id/update:
    put:
      consumes:
      - application/json
      description: |-
        Memperbarui jenis event
        * Hanya bisa diakses oleh admin
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Event Type Data
        in: body
        name: event-type
        required: true
        schema:
          $ref: '#/definitions/EventType'
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil memperbarui event type
          schema:
            $ref: '#/definitions/ResponseSuccess'
        "400":
          description: Gagal memperbarui event type
          schema:
            $ref: '#/definitions/ResponseError'
      summary: Update Event Type
      tags:
      - Event Type
  /api/event-types/create:
    post:
      consumes:
      - application/json
      description: |-
        Mendaftarkan jenis event
        * Hanya bisa diakses oleh admin
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Event Type Data
        in: body
        name: event-type
        required: true
        schema:
          $ref: '#/definitions/EventType'
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil menambahkan event type
          schema:
            $ref: '#/definitions/ResponseSuccess'
        "400":
          description: Gagal menambahkan event type
          schema:
            $ref: '#/definitions/ResponseError'
      summary: Create Event Type
      tags:
      - Event Type
  /api/event-types/list:
    get:
      consumes:
      - application/json
      description: Mendapatkan semua jenis event
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil mendapatkan event type
          schema:
            $ref: '#/definitions/ResponseSuccessWithData'
        "400":
          description: Gagal mendapatkan event type
          schema:
            $ref: '#/definitions/ResponseError'
      summary: Get All Event Type
      tags:
      - Event Type
  /api/events/:event_id:
    get:
      consumes:
      - application/json
      description: Mendapatkan detail event berdasarkan id
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil mendapatkan daftar event
          schema:
            $ref: '#/definitions/ResponseSuccessWithData'
        "400":
          description: Gagal mendapatkan daftar event
          schema:
            $ref: '#/definitions/ResponseError'
      summary: Detail Event
      tags:
      - Event
  /api/events/:event_id/delete:
    delete:
      consumes:
      - application/json
      description: |-
        Menghapus event berdasarkan id
        * Hanya bisa diakses oleh admin atau organizer
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil menghapus event
          schema:
            $ref: '#/definitions/ResponseSuccess'
        "400":
          description: Gagal menghapus event
          schema:
            $ref: '#/definitions/ResponseError'
      summary: Delete Event
      tags:
      - Event
...
  /api/events/:event_id/update:
    put:
      consumes:
      - application/json
      description: |-
        Memperbarui data event
        * Hanya bisa diakses oleh admin atau organizer
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Event Data
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/Event'
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil memperbarui event
          schema:
            $ref: '#/definitions/ResponseSuccess'
        "400":
          description: Gagal memperbarui event
          schema:
            $ref: '#/definitions/ResponseError'
      summary: Update event
      tags:
      - Event
  /api/events/create:
    post:
      consumes:
      - application/json
      description: |-
        Menambahkan data event
        * Hanya bisa diakses oleh admin atau organizer
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Event Data
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/Event'
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil menambahkan event
          schema:
            $ref: '#/definitions/ResponseSuccess'
        "400":
          description: Gagal menambahkan event
          schema:
            $ref: '#/definitions/ResponseError'
      summary: Create Event
      tags:
      - Event
  /api/events/list:
    get:
      consumes:
      - application/json
      description: Mendapatkan semua daftar event
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil mendapatkan daftar event
          schema:
            items:
              $ref: '#/definitions/ResponseSuccessWithData'
            type: array
        "400":
          description: Gagal mendapatkan daftar event
          schema:
            $ref: '#/definitions/ResponseError'
      summary: List Event
      tags:
      - Event
  /api/events/me:
    get:
      consumes:
      - application/json
      description: |-
        Mendapatkan daftar event berdasarkan id user
        * Hanya bisa diakses oleh admin atau organizer
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil mendapatkan daftar event
          schema:
            $ref: '#/definitions/ResponseSuccessWithData'
        "400":
          description: Gagal mendapatkan daftar event
          schema:
            $ref: '#/definitions/ResponseError'
      summary: Detail Event By User
      tags:
      - Event
...
swagger: "2.0"